Update and unify insights-results-aggregator-cleaner test steps
