+++ features/insights-results-aggregator-cleaner/display_old_records.feature
@@ -1,3 +1,4 @@
+@aggregator_cleaner
 Feature: Ability to display old records
 
 
@@ -43,11 +44,11 @@ Feature: Ability to display old records
       And the database is empty
      When I prepare database schema
      Then I should find that all tables are empty
-     When I insert following records into database
-          | org id | cluster name                         | timestamp  |
-          | 1      | 5d5892d4-1f74-4ccf-91af-548dfc9767aa | 2022-01-01 |
-          | 2      | 5d5892d4-1f74-4ccf-91af-548dfc9767ab | 2022-01-01 |
-          | 2      | 5d5892d4-1f74-4ccf-91af-548dfc9767ac | 2022-01-01 |
+     When I insert following records into REPORT table
+          | org_id | cluster                              | report | reported_at | last_checked_at |
+          | 1      | 5d5892d4-1f74-4ccf-91af-548dfc9767aa |        | 2930-01-01  |  2923-01-01     |
+          | 2      | 5d5892d4-1f74-4ccf-91af-548dfc9767ab |        | 2923-01-01  |  2923-01-01     |
+          | 2      | 5d5892d4-1f74-4ccf-91af-548dfc9767ac |        | 2923-01-01  |  2923-01-01     |
       And I run the cleaner to display all records older than 90 days
      Then I should see empty list of records
      When I delete all tables from database
@@ -64,14 +65,14 @@ Feature: Ability to display old records
       And the database is empty
      When I prepare database schema
      Then I should find that all tables are empty
-     When I insert following records into database
-          | org id | cluster name                         | timestamp  |
-          | 1      | 5d5892d4-1f74-4ccf-91af-548dfc9767aa | 1990-01-01 |
-          | 2      | 5d5892d4-1f74-4ccf-91af-548dfc9767ab | 1990-01-01 |
-          | 2      | 5d5892d4-1f74-4ccf-91af-548dfc9767ac | 1990-01-01 |
+     When I insert following records into REPORT table
+          | org_id | cluster                              | report | reported_at | last_checked_at |
+          | 1      | 5d5892d4-1f74-4ccf-91af-548dfc9767aa |        | 1990-01-01  |  1990-01-01     |
+          | 2      | 5d5892d4-1f74-4ccf-91af-548dfc9767ab |        | 1990-01-01  |  1990-01-01     |
+          | 2      | 5d5892d4-1f74-4ccf-91af-548dfc9767ac |        | 1990-01-01  |  1990-01-01     |
       And I run the cleaner to display all records older than 90 days
      Then I should see the following clusters
-          | cluster name                         |
+          | cluster                              |
           | 5d5892d4-1f74-4ccf-91af-548dfc9767aa |
           | 5d5892d4-1f74-4ccf-91af-548dfc9767ab |
           | 5d5892d4-1f74-4ccf-91af-548dfc9767ac |
@@ -89,15 +90,15 @@ Feature: Ability to display old records
       And the database is empty
      When I prepare database schema
      Then I should find that all tables are empty
-     When I insert following records into database
-          | org id | cluster name                         | timestamp  |
-          | 1      | 5d5892d4-1f74-4ccf-91af-548dfc9767aa | 1990-01-01 |
-          | 2      | 5d5892d4-1f74-4ccf-91af-548dfc9767ab | 1990-01-01 |
-          | 2      | 5d5892d4-1f74-4ccf-91af-548dfc9767ac | 2100-01-01 |
-          | 2      | 5d5892d4-1f74-4ccf-91af-548dfc9767ad | 2100-01-01 |
+     When I insert following records into REPORT table
+          | org_id | cluster                              | report | reported_at | last_checked_at |
+          | 1      | 5d5892d4-1f74-4ccf-91af-548dfc9767aa |        | 1990-01-01  |  1990-01-01     |
+          | 2      | 5d5892d4-1f74-4ccf-91af-548dfc9767ab |        | 1990-01-01  |  1990-01-01     |
+          | 2      | 5d5892d4-1f74-4ccf-91af-548dfc9767ac |        | 2900-01-01  |  2200-01-01     |
+          | 2      | 5d5892d4-1f74-4ccf-91af-548dfc9767ad |        | 2900-01-01  |  2200-01-01     |
       And I run the cleaner to display all records older than 90 days
      Then I should see the following clusters
-          | cluster name                         |
+          | cluster                              |
           | 5d5892d4-1f74-4ccf-91af-548dfc9767aa |
           | 5d5892d4-1f74-4ccf-91af-548dfc9767ab |
      When I delete all tables from database
