+++ features/insights-results-aggregator-cleaner/smoketests.feature
@@ -1,3 +1,4 @@
+@aggregator_cleaner
 Feature: Basic set of smoke tests
 
 
@@ -34,18 +35,19 @@ Feature: Basic set of smoke tests
      Then I should be disconnected
 
 
-  @database
-  Scenario Outline: Check if the test database does not contain tables to be created by tests
-    Given the system is in default state
-     When I connect to database named test as user postgres with password postgres
-     Then I should be able to connect to such database
-     When I look for the table <table> in database
-     Then I should not be able to find it
-
-        Examples: tables in database
-            | table                              |
-            | report                             |
-            | cluster_rule_toggle                |
-            | cluster_rule_user_feedback         |
-            | cluster_user_rule_disable_feedback |
-            | rule_hit                           |
+# I think we don't need this test
+#  @database0
+#  Scenario Outline: Check if the test database does not contain tables to be created by tests
+#    Given the system is in default state
+#     When I connect to database named test as user postgres with password postgres
+#     Then I should be able to connect to such database
+#     When I look for the table <table> in database
+#     Then I should not be able to find it
+#
+#        Examples: tables in database
+#            | table                              |
+#            | report                             |
+#            | cluster     _rule_toggle                |
+#            | cluster     _rule_user_feedback         |
+#            | cluster     _user_rule_disable_feedback |
+#            | rule_hit                           |
