+++ features/steps/cleaner_main.py
@@ -17,6 +17,9 @@
 import subprocess
 from src.process_output import process_generated_output
 
+from behave import then, when
+
+
 # default name of file generated by Insights Aggregator Cleaner during testing
 test_output = "test"
 
@@ -68,7 +71,7 @@ def run_cleaner_to_cleanup_cluster(context, cluster):
 
 def check_help_from_cleaner(context):
     """Check if help is displayed by cleaner."""
-    expected_output = """
+    expected_output = """Clowder is not enabled, skipping init...
 Usage of insights-results-aggregator-cleaner:
   -authors
         show authors
@@ -84,8 +87,12 @@ Usage of insights-results-aggregator-cleaner:
         list clusters with the same rule(s) disabled by different users
   -output string
         filename for old cluster listing
+  -show-configuration
+        show configuration
   -summary
         print summary table after cleanup
+  -vacuum
+        vacuum database
   -version
         show cleaner version"""
 
@@ -138,7 +145,7 @@ def check_empty_list_of_records(context):
 def check_non_empty_list_of_records(context):
     """Check if the cleaner displays the suggested clusters."""
     # set of expected clusters
-    expected_clusters = set(item["cluster name"] for item in context.table)
+    expected_clusters = set(item["cluster"] for item in context.table)
 
     # set of actually found clusters
     found_clusters = set()
