+++ features/steps/cleaner_storage.py
@@ -13,169 +13,11 @@
 # limitations under the License.
 
 """Database-related operations performed by BDD tests."""
-
 import psycopg2
 from psycopg2.errors import UndefinedTable
 from src.sql import construct_insert_statement
-
 from behave import given, then, when
-
-
-CREATE_TABLE_ADVISOR_RATINGS = """
-CREATE TABLE advisor_ratings (
-                                user_id         VARCHAR NOT NULL,
-                                org_id          VARCHAR NOT NULL,
-                                rule_fqdn       VARCHAR NOT NULL,
-                                error_key       VARCHAR NOT NULL,
-                                rated_at        TIMESTAMP,
-                                last_updated_at TIMESTAMP,
-                                rating          SMALLINT,
-                                rule_id         VARCHAR NOT NULL
-                        );
-"""
-
-CREATE_TABLE_REPORT = """
-CREATE TABLE report (
-                                org_id          INTEGER NOT NULL,
-                                cluster         VARCHAR NOT NULL UNIQUE,
-                                report          VARCHAR NOT NULL,
-                                reported_at     TIMESTAMP,
-                                last_checked_at TIMESTAMP,
-                                kafka_offset    BIGINT NOT NULL DEFAULT 0,
-                                gathered_at     TIMESTAMP,
-                                PRIMARY KEY(org_id, cluster)
-                        );
-
-"""
-
-CREATE_TABLE_CLUSTER_RULE_TOGGLE = """
-CREATE TABLE cluster_rule_toggle (
-                                cluster_id  VARCHAR NOT NULL,
-                                rule_id     VARCHAR NOT NULL,
-                                user_id     VARCHAR NOT NULL,
-                                disabled    SMALLINT NOT NULL,
-                                disabled_at TIMESTAMP NULL,
-                                enabled_at  TIMESTAMP NULL,
-                                updated_at  TIMESTAMP NOT NULL,
-                                error_key   VARCHAR NOT NULL
-                        );
-"""
-
-CREATE_TABLE_CLUSTER_RULE_USER_FEEDBACK = """
-CREATE TABLE cluster_rule_user_feedback (
-                                        cluster_id VARCHAR NOT NULL,
-                                        rule_id    VARCHAR NOT NULL,
-                                        user_id    VARCHAR NOT NULL,
-                                        message    VARCHAR NOT NULL,
-                                        user_vote  SMALLINT NOT NULL,
-                                        added_at   TIMESTAMP NOT NULL,
-                                        updated_at TIMESTAMP NOT NULL,
-                                        error_key  VARCHAR NOT NULL
-                        );
-"""
-
-CREATE_TABLE_CLUSTER_USER_RULE_DISABLE_FEEDBACK = """
-CREATE TABLE cluster_user_rule_disable_feedback (
-                                cluster_id VARCHAR NOT NULL,
-                                user_id    VARCHAR NOT NULL,
-                                rule_id    VARCHAR NOT NULL,
-                                message    VARCHAR NOT NULL,
-                                added_at   TIMESTAMP NOT NULL,
-                                updated_at TIMESTAMP NOT NULL,
-                                error_key  VARCHAR NOT NULL
-                        );
-"""
-
-CREATE_TABLE_RULE_HIT = """
-CREATE TABLE rule_hit (
-                        org_id          INTEGER NOT NULL,
-                        cluster_id      VARCHAR NOT NULL,
-                        rule_fqdn       VARCHAR NOT NULL,
-                        error_key       VARCHAR NOT NULL,
-                        template_data   VARCHAR NOT NULL,
-                        PRIMARY KEY(cluster_id, org_id, rule_fqdn, error_key)
-                        );
-"""
-
-CREATE_TABLE_CONSUMER_ERROR = """
-CREATE TABLE consumer_error (
-                        topic        VARCHAR NOT NULL,
-                        partition    INTEGER NOT NULL,
-                        topic_offset INTEGER NOT NULL,
-                        key          VARCHAR,
-                        produced_at  TIMESTAMP NOT NULL,
-                        consumed_at  TIMESTAMP NOT NULL,
-                        message      VARCHAR,
-                        error        VARCHAR NOT NULL
-                        );
-"""
-
-CREATE_TABLE_MIGRATION_INFO = """
-CREATE TABLE migration_info (
-                        version INTEGER NOT NULL
-);
-"""
-
-CREATE_TABLE_RECOMMENDATION = """
-CREATE TABLE recommendation (
-                        org_id     INTEGER NOT NULL,
-                        cluster_id VARCHAR NOT NULL,
-                        rule_fqdn  TEXT NOT NULL,
-                        error_key  VARCHAR NOT NULL,
-                        rule_id    VARCHAR NOT NULL,
-                        created_at TIMESTAMP
-);
-"""
-
-CREATE_TABLE_REPORT_INFO = """
-CREATE TABLE report_info (
-                        org_id       INTEGER NOT NULL,
-                        cluster_id   VARCHAR NOT NULL,
-                        version_info VARCHAR NOT NULL
-);
-"""
-
-CREATE_TABLE_RULE_DISABLE = """
-CREATE TABLE rule_disable (
-                        org_id        VARCHAR NOT NULL,
-                        user_id       VARCHAR NOT NULL,
-                        rule_id       VARCHAR NOT NULL,
-                        error_key     VARCHAR NOT NULL,
-                        justification VARCHAR,
-                        created_at    TIMESTAMP,
-                        updated_at    TIMESTAMP
-);
-"""
-
-# all commands to create tables
-CREATE_TABLE_COMMANDS = (
-    CREATE_TABLE_ADVISOR_RATINGS,
-    CREATE_TABLE_REPORT,
-    CREATE_TABLE_CLUSTER_RULE_TOGGLE,
-    CREATE_TABLE_CLUSTER_RULE_USER_FEEDBACK,
-    CREATE_TABLE_CLUSTER_USER_RULE_DISABLE_FEEDBACK,
-    CREATE_TABLE_RULE_HIT,
-    CREATE_TABLE_CONSUMER_ERROR,
-    CREATE_TABLE_MIGRATION_INFO,
-    CREATE_TABLE_RECOMMENDATION,
-    CREATE_TABLE_REPORT_INFO,
-    CREATE_TABLE_RULE_DISABLE,
-)
-
-# following tables should be processed
-DB_TABLES = (
-    "advisor_ratings",
-    "report",
-    "cluster_rule_toggle",
-    "cluster_rule_user_feedback",
-    "cluster_user_rule_disable_feedback",
-    "rule_hit",
-    "consumer_error",
-    "migration_info",
-    "recommendation",
-    "report_info",
-    "rule_disable",
-)
+from common_aggregator import DB_TABLES
 
 
 @when(u"I look for the table {table} in database")
@@ -184,9 +26,9 @@ def look_for_table(context, table):
     cursor = context.connection.cursor()
     try:
         cursor.execute("SELECT 1 from {}".format(table))
-        v = cursor.fetchone()
+        _ = cursor.fetchone()
         context.table_found = True
-    except UndefinedTable as e:
+    except UndefinedTable:
         context.table_found = False
 
     context.connection.commit()
@@ -226,8 +68,14 @@ def establish_connection_to_database(context):
     assert context.database_name is not None
     assert context.database_user is not None
     assert context.database_password is not None
-    connection_string = "dbname={} user={} password={}".format(
-        context.database_name, context.database_user, context.database_password
+    assert context.database_host is not None
+    assert context.database_port is not None
+    connection_string = "host={} port={} dbname={} user={} password={}".format(
+        context.database_host,
+        context.database_port,
+        context.database_name,
+        context.database_user,
+        context.database_password,
     )
     context.connection = psycopg2.connect(connection_string)
     assert context.connection is not None, "connection should be established"
@@ -242,13 +90,13 @@ def ensure_database_emptiness(context):
     for table in DB_TABLES:
         try:
             cursor.execute("SELECT 1 from {}".format(table))
-            v = cursor.fetchone()
+            _ = cursor.fetchone()
             context.connection.commit()
+            print("DB name: ", context.connection.info.dsn_parameters)
             raise Exception("Table '{}' exists".format(table))
-        except UndefinedTable as e:
+        except UndefinedTable:
             # exception means that the table does not exists
             context.connection.rollback()
-            pass
 
 
 @then(u"I should find that all tables are empty")
@@ -263,21 +111,8 @@ def ensure_data_tables_emptiness(context):
                 len(results)
             )
             assert results[0] == 0, "Table '{}' is not empty as expected".format(table)
-        except Exception as e:
-            raise e
-
-
-@when(u"I prepare database schema")
-def prepare_database_schema(context):
-    """Prepare database schema."""
-    cursor = context.connection.cursor()
-    try:
-        for createTableCommand in CREATE_TABLE_COMMANDS:
-            cursor.execute(createTableCommand)
-            context.connection.commit()
-    except Exception as e:
-        context.connection.rollback()
-        raise e
+        except Exception:
+            raise
 
 
 @when(u"I delete all tables from database")
@@ -288,9 +123,9 @@ def delete_all_tables(context):
         try:
             cursor.execute("DROP TABLE {}".format(table))
             context.connection.commit()
-        except Exception as e:
+        except Exception:
             context.connection.rollback()
-            raise e
+            raise
 
 
 @when(u"I insert following records into {table} table")
@@ -307,9 +142,10 @@ def insert_records_into_selected_table(context, table):
         # perform several INSERTs
         for row in context.table:
             # try to perform INSERT statement
+            print(row)
             cursor.execute(insert_statement, row)
 
         context.connection.commit()
-    except Exception as e:
+    except Exception:
         context.connection.rollback()
-        raise e
+        raise
