+++ features/steps/common_db.py
@@ -14,24 +14,23 @@
 
 """Common steps for DB related operations."""
 
-from behave import then, when, given
 
-import subprocess
 import psycopg2
+from behave import when, then
 
 
 @when(u"I connect to database named {database} as user {user} with password {password}")
-def connect_to_database(context, database, user, password, host="localhost"):
+def connect_to_database(context, database, user, password):
     """Perform connection to selected database."""
-    connection_string = "dbname={} user={} password={} host={}".format(
-        database, user, password, host
+    connection_string = "dbname={} user={} password={} host={} port={}".format(
+        database, user, password, context.database_host, context.database_port
     )
     context.connection = psycopg2.connect(connection_string)
 
 
 @then(u"I should be able to connect to such database")
 def check_connection(context):
-    """Chck the connection to database."""
+    """Check the connection to database."""
     assert context.connection is not None, "connection should be established"
 
 
@@ -46,20 +45,3 @@ def disconnect_from_database(context):
 def check_disconnection(context):
     """Check that the connection has been closed."""
     assert context.connection is None, "connection should be closed"
-
-
-@given("Postgres is running")
-def check_if_postgres_is_running(context):
-    """Check if Postgresql service is active."""
-    p = subprocess.Popen(["pg_isready", "-q", "--host", "localhost", "--port", "5432"])
-    assert p is not None
-
-    # interact with the process:
-    p.communicate()
-
-    # check the return code
-    assert (
-        p.returncode == 0
-    ), "Postgresql service not running: got return code {code}".format(
-        code=p.returncode
-    )
